wip: basic cli

+++ src/setupTests.js
@@ -7,7 +7,14 @@ import { suggestionSettings } from "./constants";
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 
 /* eslint-disable functional/immutable-data */
-global.crypto.subtle = crypto.webcrypto.subtle;
+if (!global.crypto) {
+  global.crypto = crypto.webcrypto;
+} else if (!global.crypto.subtle) {
+  Object.defineProperty(global.crypto, "subtle", {
+    configurable: true,
+    value: crypto.webcrypto.subtle,
+  });
+}
 global.TextEncoder = TextEncoder;
 global.TextDecoder = TextDecoder;
 global.Buffer = Buffer;
@@ -15,11 +22,17 @@ global.ResizeObserver = function ResizeObserver() {};
 global.ResizeObserver.prototype.observe = () => {};
 global.ResizeObserver.prototype.unobserve = () => {};
 global.ResizeObserver.prototype.disconnect = () => {};
-Element.prototype.scrollIntoView = () => {};
+if (typeof Element !== "undefined") {
+  Element.prototype.scrollIntoView = () => {};
+}
 suggestionSettings.maxSuggestions = 3;
 /* eslint-enable functional/immutable-data */
 
 afterEach(() => {
-  localStorage.clear();
-  window.history.pushState({}, "", "/");
+  if (typeof localStorage !== "undefined") {
+    localStorage.clear();
+  }
+  if (typeof window !== "undefined") {
+    window.history.pushState({}, "", "/");
+  }
 });
